+++ app/Config/LinkMap.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Config;
+
+class LinkMap
+{
+    public $map = [
+        'example' => ['dashboard/example', false],
+        'git' => ['https://github.com/I-ara-I/php-framework', true],
+    ];
+}