+++ app/View/index.php
@@ -1,1 +1,5 @@
 <p><?= $msg ?></p>
+
+<a href="<?= $this->getLink('example') ?>">Example</a>
+<a href="<?= $this->createLink('dashboard') ?>">dashboard</a>
+<a href="<?= $this->getLink('git', false) ?>">GitHub</a>